Started splitting up LGSG because it's gotten so large
